pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Bezerra <bezerra.fernandes69@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.1"
tensorflow-datasets = "^4.8.0"
pillow = "^9.4.0"
matplotlib = "^3.6.2"
opencv-python = "^4.7.0.68"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
